# Conflicts:
#	README.md
#	contrib/Orochi
#	tools/premake5/linux64/premake5
#	tutorials/01_geom_intersection/main.cpp
#	tutorials/02_scene_intersection/main.cpp
#	tutorials/03_custom_intersection/TestKernel.h
#	tutorials/03_custom_intersection/main.cpp
#	tutorials/04_shared_stack/TestKernel.h
#	tutorials/05_custom_bvh_import/main.cpp
#	tutorials/06_obj_AO/TestKernel.h
#	tutorials/07_motion_blur/main.cpp
#	tutorials/common/TestBase.cpp
#	tutorials/common/TestBase.h
#	tutorials/premake5.lua

+++ tutorials/common/TestBase.h
@@ -278,7 +278,7 @@ class TestBase
 	};
 
 	template <typename T>
-	void dMalloc( T*& ptr, size_t n )
+	void dMalloc( T*& ptr, int n )
 	{
 		oroError e = oroMalloc( (oroDeviceptr*)&ptr, sizeof( T ) * n );
 		ASSERT( e == oroSuccess );
@@ -290,21 +290,21 @@ class TestBase
 		ASSERT( e == oroSuccess );
 	}
 
-	void dMemset( void* ptr, int val, size_t n )
+	void dMemset( void* ptr, int val, int n )
 	{
 		oroError e = oroMemset( (oroDeviceptr)ptr, val, n );
 		ASSERT( e == oroSuccess );
 	}
 
 	template <typename T>
-	void dCopyHtoD( T* dst, T* src, size_t n )
+	void dCopyHtoD( T* dst, T* src, int n )
 	{
 		oroError e = oroMemcpyHtoD( (oroDeviceptr)dst, src, sizeof( T ) * n );
 		ASSERT( e == oroSuccess );
 	}
 
 	template <typename T>
-	void dCopyDtoH( T* dst, T* src, size_t n )
+	void dCopyDtoH( T* dst, T* src, int n )
 	{
 		oroError e = oroMemcpyDtoH( dst, (oroDeviceptr)src, sizeof( T ) * n );
 		ASSERT( e == oroSuccess );
